Add benchmark + train entrypoint
