+++ src/anomalib/config/helpers.py
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from pathlib import Path, PosixPath
-from typing import Type
+from typing import Type, cast
 
 import yaml
 from jsonargparse import Namespace
@@ -52,3 +52,32 @@ def to_yaml(config: Namespace | ListConfig | DictConfig) -> str:
         _config = OmegaConf.create(_config.as_dict())
     container = OmegaConf.to_container(_config, enum_to_str=True)
     return yaml.dump(container, default_flow_style=False, allow_unicode=True, sort_keys=False, Dumper=_get_dumper())
+
+
+def to_tuple(input_size: int | ListConfig) -> tuple[int, int]:
+    """Converts int or list to a tuple.
+
+    Example:
+        >>> to_tuple(256)
+        (256, 256)
+        >>> to_tuple([256, 256])
+        (256, 256)
+
+    Args:
+        input_size (int | ListConfig): input_size
+
+    Raises:
+        ValueError: Unsupported value type.
+
+    Returns:
+        tuple[int, int]: Tuple of input_size
+    """
+    ret_val: tuple[int, int]
+    if isinstance(input_size, int):
+        ret_val = cast(tuple[int, int], (input_size,) * 2)
+    elif isinstance(input_size, ListConfig):
+        assert len(input_size) == 2, "Expected a single integer or tuple of length 2 for width and height."
+        ret_val = cast(tuple[int, int], tuple(input_size))
+    else:
+        raise ValueError(f"Expected either int or ListConfig, got {type(input_size)}")
+    return ret_val
