+++ src/anomalib/utils/hpo/sweep.py
@@ -1,0 +1,100 @@
+"""Utilities to run hyperparameter optimization."""
+
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from argparse import ArgumentParser
+from enum import Enum
+from pathlib import Path
+from typing import Optional, Union
+
+from lightning.pytorch import seed_everything
+from lightning.pytorch.cli import LightningArgumentParser
+from omegaconf import OmegaConf
+
+from anomalib.config import get_configurable_parameters
+
+from .runners import CometSweep, WandbSweep
+
+
+class HPOBackend(str, Enum):
+    """HPO backend choices."""
+
+    COMET = "comet"
+    WANDB = "wandb"
+
+    def __str__(self):
+        """Return the string representation of the enum value."""
+        return self.value
+
+
+def get_hpo_parser(parser: Optional[Union[ArgumentParser, LightningArgumentParser]] = None):
+    """Gets the HPO parser."""
+    if parser is None:
+        parser = ArgumentParser()
+    parser.add_argument(
+        "--model", type=str, default="padim", help="Name of the algorithm to train/test", required=False
+    )
+    parser.add_argument("--model_config", type=Path, required=False, help="Path to a model config file")
+    parser.add_argument("--sweep_config", type=Path, required=True, help="Path to sweep configuration")
+    parser.add_argument(
+        "--backend", type=HPOBackend, default=HPOBackend.WANDB, help="HPO backend to use", required=False
+    )
+    parser.add_argument(
+        "--entity",
+        type=str,
+        required=False,
+        help="Username or workspace where you want to send your runs to. If not set, the default workspace is used.",
+    )
+
+    return parser
+
+
+class Sweep:
+    """HPO class to run hyperparameter optimization.
+
+    Args:
+        model (str): Name of the algorithm to train/test. If not provided, the model name is read from the model config.
+        model_config (Optional[Union[Path, str]]): Path to a model config file. If not provided, the model is loaded
+            based on the model name.
+        sweep_config (Optional[Union[Path, str]]): Path to sweep configuration. The configuration depends on the type
+            of backend.
+        backend (HPOBackend): HPO backend to use. Defaults to ```Backend.COMET```.
+        entity (Optional[str]): Username or workspace where you want to send your runs to. If not set, the default
+            workspace is used.
+    """
+
+    def __init__(
+        self,
+        model: Optional[str],
+        model_config: Optional[Union[Path, str]],
+        sweep_config: Union[Path, str],
+        backend: HPOBackend = HPOBackend.COMET,
+        entity: Optional[str] = "",
+    ):
+        if model is None and model_config is None:
+            raise ValueError("Either model or model_config must be provided.")
+
+        self.model_config = get_configurable_parameters(model_name=model, config_path=model_config)
+        self.sweep_config = OmegaConf.load(sweep_config)
+        self.entity = entity
+
+        self.runner = self.get_runner(backend)
+
+    def get_runner(self, backend: HPOBackend) -> Union[CometSweep, WandbSweep]:
+        """Gets the runner for the specified backend."""
+        runner: Union[CometSweep, WandbSweep]
+        if backend == HPOBackend.COMET:
+            runner = CometSweep(self.model_config, self.sweep_config, self.entity)
+        elif backend == HPOBackend.WANDB:
+            runner = WandbSweep(self.model_config, self.sweep_config, self.entity)
+        else:
+            raise ValueError(f"Unknown backend {backend}")
+        return runner
+
+    def run(self):
+        """Runs the sweep."""
+        if self.model_config.get("seed_everything") is not None:
+            seed_everything(self.model_config.seed_everything)
+
+        self.runner.run()
