--- tools/benchmarking/utils/__init__.py
@@ -1,8 +1,0 @@
-"""Utils specific to running benchmarking scripts."""
-
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .metrics import upload_to_comet, upload_to_wandb, write_metrics
-
-__all__ = ["write_metrics", "upload_to_comet", "upload_to_wandb"]
